cloud_functions/get_description/requirements.txt
functions-framework==3.*
Flask
python-dotenv==1.0.0
openai==1.6.1
firebase_admin==6.3.0